requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0
django-browser-reload==1.12.1
django-crispy-forms==2.1
django-import-export==3.3.5
et-xmlfile==1.1.0
MarkupPy==1.14
mysqlclient==2.2.1
mysqlparse==0.1.6
odfpy==1.4.1
openpyxl==3.1.2
pyparsing==3.1.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.4
xlrd==2.0.1
xlwt==1.3.0
